+++ lib/f_service/result/base.rb
@@ -7,6 +7,8 @@ module FService
     #
     # @abstract
     class Base
+      attr_reader :types
+
       %i[and_then successful? failed? value value! error].each do |method_name|
         define_method(method_name) do |*_args|
           raise NotImplementedError, "called #{method_name} on class Result::Base"
@@ -14,8 +16,17 @@ module FService
       end
 
       # You usually shouldn't call this directly. See {FService::Base#Failure} and {FService::Base#Success}.
-      def initialize
+      def initialize(types = [])
         @handled = false
+        @types = types
+        @matching_types = []
+      end
+
+      # Implements old attribute type. Its deprecated in favor of using types.
+      def type
+        FService.deprecate!(name: "#{self.class}##{__method__}", alternative: '#types', from: caller[0])
+
+        types.size == 1 ? types.first : Array(@matching_types).first
       end
 
       # This hook runs if the result is successful.
@@ -59,6 +70,7 @@ module FService
       # @api public
       def on_success(*target_types, unhandled: false)
         if successful? && unhandled? && expected_type?(target_types, unhandled: unhandled)
+          match_types(target_types)
           yield(*to_ary)
           @handled = true
           freeze
@@ -108,6 +120,7 @@ module FService
       # @api public
       def on_failure(*target_types, unhandled: false)
         if failed? && unhandled? && expected_type?(target_types, unhandled: unhandled)
+          match_types(target_types)
           yield(*to_ary)
           @handled = true
           freeze
@@ -122,7 +135,7 @@ module FService
       def to_ary
         data = successful? ? value : error
 
-        [data, type]
+        [data, @matching_types.first]
       end
 
       private
@@ -136,7 +149,11 @@ module FService
       end
 
       def expected_type?(target_types, unhandled:)
-        target_types.empty? || unhandled || target_types.include?(type)
+        target_types.empty? || unhandled || target_types.any? { |target_type| types.include?(target_type) }
+      end
+
+      def match_types(target_types)
+        @matching_types = target_types.empty? ? types : target_types & types
       end
     end
   end