+++ spec/f_service/base_spec.rb
@@ -12,7 +12,7 @@ RSpec.describe FService::Base do
   describe '#Success' do
     subject(:response) { described_class.new.Success(:ok, data: 'yay!') }
 
-    it { expect(response.type).to eq(:ok) }
+    it { expect(response.types).to contain_exactly(:ok) }
     it { expect(response.value).to eq('yay!') }
   end
 
@@ -25,7 +25,7 @@ RSpec.describe FService::Base do
   describe '#Failure' do
     subject(:response) { described_class.new.Failure(:error, data: 'Whoops!') }
 
-    it { expect(response.type).to eq(:error) }
+    it { expect(response.types).to contain_exactly(:error) }
     it { expect(response.error).to eq('Whoops!') }
   end
 
@@ -34,31 +34,31 @@ RSpec.describe FService::Base do
       subject(:response) { described_class.new.Check(:math_works) { 1 < 2 } }
 
       it { expect(response).to be_successful }
-      it { expect(response.type).to eq(:math_works) }
-      it { expect(response.value!).to eq(true) }
+      it { expect(response.types).to contain_exactly(:math_works) }
+      it { expect(response.value!).to be_truthy }
     end
 
     context 'when block evaluates to false' do
       subject(:response) { described_class.new.Check(:math_works) { 1 > 2 } }
 
       it { expect(response).to be_failed }
-      it { expect(response.type).to eq(:math_works) }
-      it { expect(response.error).to eq(false) }
+      it { expect(response.types).to contain_exactly(:math_works) }
+      it { expect(response.error).to be_falsy }
     end
 
     context 'when type is not specified' do
       subject(:response) { described_class.new.Check { 1 > 2 } }
 
       it { expect(response).to be_failed }
-      it { expect(response.type).to eq(nil) }
-      it { expect(response.error).to eq(false) }
+      it { expect(response.types).to be_empty }
+      it { expect(response.error).to be_falsy }
     end
 
     context 'when data is passed' do
       subject(:response) { described_class.new.Check(data: 'that is an error') { 1 > 2 } }
 
       it { expect(response).to be_failed }
-      it { expect(response.type).to eq(nil) }
+      it { expect(response.types).to be_empty }
       it { expect(response.error).to eq('that is an error') }
     end
   end
@@ -67,14 +67,14 @@ RSpec.describe FService::Base do
     subject(:response) { described_class.new.Try(:division) { 0 / 1 } }
 
     it { expect(response).to be_successful }
-    it { expect(response.type).to eq(:division) }
-    it { expect(response.value!).to eq(0) }
+    it { expect(response.types).to contain_exactly(:division) }
+    it { expect(response.value!).to be_zero }
 
     context 'when some exception is raised' do
       subject(:response) { described_class.new.Try(:division) { 1 / 0 } }
 
       it { expect(response).to be_failed }
-      it { expect(response.type).to eq(:division) }
+      it { expect(response.types).to contain_exactly(:division) }
       it { expect(response.error).to be_a ZeroDivisionError }
     end
 
@@ -82,8 +82,8 @@ RSpec.describe FService::Base do
       subject(:response) { described_class.new.Try { 0 / 1 } }
 
       it { expect(response).to be_successful }
-      it { expect(response.type).to eq(nil) }
-      it { expect(response.value!).to eq 0 }
+      it { expect(response.types).to be_empty }
+      it { expect(response.value!).to be_zero }
     end
 
     context 'when raised exception does not match specified exception' do